Added update class and minor bug fixes
Added a class for checking if the plugin is up-to-date. Since this
plugin is not on wordpress.org there was no way to keep track of
updates. This class will now display a message if the plugin is out of
date if it detects the plugin is out of date (checks every 6 hours).

- updated plugin version in dpsfa.php to be in the same convention as
git releases
- took out forced error reporting in dpsfa.php
- fixed all hardcoded references to the plugin folder