Bug Fix: Image fallback when bundling article
If images inside an article link to a smaller image size, Wordpress may
not be able to find the original image. A fallback was added in these
cases to try to find the original image and in cases where the image
isn’t in the media library the plugin will download the image and
attach it to the article.