Bug fix: Unable to upload
Resolved a bug that was passing the folio as the value instead of a
blank array in the apply_filters() function.

apply_filters is a function that allows individual templates to pass
additional files while the article is being bundled. For example,
slideshows or any other images/files that the article needs.