Bug Fixes and Article Download
This update resolves a bug in the template engine that caused previews
from breaking.

Also added a new feature to download article bundles. On the article
page, there’s a “Download Archive” button in the actions box that will
download a zip of the article that gets uploaded to DPS. This archive
can be used to diagnose theme problems or can be uploaded directly into
InDesign’s folio panel as an HTML Article.